+++ src/main/java/org/jetbrains/research/kotlinrminer/decomposition/AbstractCodeFragment.java
@@ -96,7 +96,7 @@ public abstract class AbstractCodeFragment implements LocationInfoProvider {
                         }
                         String beforeMatch = afterReplacements.substring(0, start);
                         String afterMatch =
-                                afterReplacements.substring(start + parameter.length(), afterReplacements.length());
+                                afterReplacements.substring(start + parameter.length());
                         if (quoteBefore(beforeMatch) && quoteAfter(afterMatch)) {
                             isInsideStringLiteral = true;
                         }
@@ -228,7 +228,7 @@ public abstract class AbstractCodeFragment implements LocationInfoProvider {
             if (initializer.startsWith("(")) {
                 //ignore casting
                 String initializerWithoutCasting =
-                        initializer.substring(initializer.indexOf(")") + 1, initializer.length());
+                        initializer.substring(initializer.indexOf(")") + 1);
                 return initializerWithoutCasting.equals(expression);
             }
         }
