+++ src/main/java/org/jetbrains/research/kotlinrminer/decomposition/AbstractExpression.java
@@ -74,7 +74,7 @@ public class AbstractExpression extends AbstractCodeFragment {
     }
 
     public String toString() {
-        return getExpression().toString();
+        return getExpression();
     }
 
     @Override
