+++ src/main/java/org/jetbrains/research/kotlinrminer/decomposition/UMLOperationBodyMapper.java
@@ -5,6 +5,9 @@ import org.jetbrains.research.kotlinrminer.api.Refactoring;
 import org.jetbrains.research.kotlinrminer.api.RefactoringMinerTimedOutException;
 import org.jetbrains.research.kotlinrminer.decomposition.replacement.*;
 import org.jetbrains.research.kotlinrminer.diff.*;
+import org.jetbrains.research.kotlinrminer.diff.refactoring.CandidateAttributeRefactoring;
+import org.jetbrains.research.kotlinrminer.diff.refactoring.CandidateMergeVariableRefactoring;
+import org.jetbrains.research.kotlinrminer.diff.refactoring.CandidateSplitVariableRefactoring;
 import org.jetbrains.research.kotlinrminer.uml.UMLAttribute;
 import org.jetbrains.research.kotlinrminer.uml.UMLOperation;
 import org.jetbrains.research.kotlinrminer.uml.UMLParameter;
