+++ src/main/java/org/jetbrains/research/kotlinrminer/decomposition/VariableReplacementAnalysis.java
@@ -18,16 +18,16 @@ import org.jetbrains.research.kotlinrminer.decomposition.replacement.Replacement
 import org.jetbrains.research.kotlinrminer.decomposition.replacement.SplitVariableReplacement;
 import org.jetbrains.research.kotlinrminer.decomposition.replacement.VariableDeclarationReplacement;
 import org.jetbrains.research.kotlinrminer.decomposition.replacement.VariableReplacementWithMethodInvocation;
-import org.jetbrains.research.kotlinrminer.diff.CandidateAttributeRefactoring;
-import org.jetbrains.research.kotlinrminer.diff.CandidateMergeVariableRefactoring;
-import org.jetbrains.research.kotlinrminer.diff.CandidateSplitVariableRefactoring;
-import org.jetbrains.research.kotlinrminer.diff.ChangeVariableTypeRefactoring;
-import org.jetbrains.research.kotlinrminer.diff.ExtractAttributeRefactoring;
-import org.jetbrains.research.kotlinrminer.diff.ExtractVariableRefactoring;
-import org.jetbrains.research.kotlinrminer.diff.InlineVariableRefactoring;
-import org.jetbrains.research.kotlinrminer.diff.MergeVariableRefactoring;
-import org.jetbrains.research.kotlinrminer.diff.RenameVariableRefactoring;
-import org.jetbrains.research.kotlinrminer.diff.SplitVariableRefactoring;
+import org.jetbrains.research.kotlinrminer.diff.refactoring.CandidateAttributeRefactoring;
+import org.jetbrains.research.kotlinrminer.diff.refactoring.CandidateMergeVariableRefactoring;
+import org.jetbrains.research.kotlinrminer.diff.refactoring.CandidateSplitVariableRefactoring;
+import org.jetbrains.research.kotlinrminer.diff.refactoring.ChangeVariableTypeRefactoring;
+import org.jetbrains.research.kotlinrminer.diff.refactoring.ExtractAttributeRefactoring;
+import org.jetbrains.research.kotlinrminer.diff.refactoring.ExtractVariableRefactoring;
+import org.jetbrains.research.kotlinrminer.diff.refactoring.InlineVariableRefactoring;
+import org.jetbrains.research.kotlinrminer.diff.refactoring.MergeVariableRefactoring;
+import org.jetbrains.research.kotlinrminer.diff.refactoring.RenameVariableRefactoring;
+import org.jetbrains.research.kotlinrminer.diff.refactoring.SplitVariableRefactoring;
 import org.jetbrains.research.kotlinrminer.diff.UMLClassBaseDiff;
 import org.jetbrains.research.kotlinrminer.diff.UMLOperationDiff;
 import org.jetbrains.research.kotlinrminer.diff.UMLParameterDiff;
