+++ src/main/java/org/jetbrains/research/kotlinrminer/diff/RenamePattern.java
@@ -1,8 +1,8 @@
 package org.jetbrains.research.kotlinrminer.diff;
 
 public class RenamePattern {
-    private String before;
-    private String after;
+    private final String before;
+    private final String after;
 
     public RenamePattern(String originalPath, String movedPath) {
         this.before = originalPath;
