+++ src/main/java/org/jetbrains/research/kotlinrminer/diff/UMLClassMoveDiff.java
@@ -13,13 +13,11 @@ public class UMLClassMoveDiff extends UMLClassBaseDiff {
     }
 
     public String toString() {
-        StringBuilder sb = new StringBuilder();
-        sb.append("class ");
-        sb.append(originalClass.getName());
-        sb.append(" was moved to ");
-        sb.append(nextClass.getName());
-        sb.append("\n");
-        return sb.toString();
+        return "class " +
+                originalClass.getName() +
+                " was moved to " +
+                nextClass.getName() +
+                "\n";
     }
 
     public boolean equals(Object o) {
@@ -29,7 +27,8 @@ public class UMLClassMoveDiff extends UMLClassBaseDiff {
 
         if (o instanceof UMLClassMoveDiff) {
             UMLClassMoveDiff classMoveDiff = (UMLClassMoveDiff) o;
-            return this.originalClass.equals(classMoveDiff.originalClass) && this.nextClass.equals(classMoveDiff.nextClass);
+            return this.originalClass.equals(classMoveDiff.originalClass) && this.nextClass.equals(
+                    classMoveDiff.nextClass);
         }
         return false;
     }
