+++ src/main/java/org/jetbrains/research/kotlinrminer/diff/UMLClassRenameDiff.java
@@ -16,12 +16,10 @@ public class UMLClassRenameDiff extends UMLClassBaseDiff {
     }
 
     public String toString() {
-        StringBuilder sb = new StringBuilder();
-        sb.append("class ");
-        sb.append(originalClass.getName());
-        sb.append(" was renamed to ");
-        sb.append(nextClass.getName());
-        sb.append("\n");
-        return sb.toString();
+        return "class " +
+                originalClass.getName() +
+                " was renamed to " +
+                nextClass.getName() +
+                "\n";
     }
 }
