+++ src/main/java/org/jetbrains/research/kotlinrminer/diff/refactoring/RenameVariableRefactoring.java
@@ -1,0 +1,156 @@
+package org.jetbrains.research.kotlinrminer.diff.refactoring;
+
+import java.util.ArrayList;
+import java.util.LinkedHashSet;
+import java.util.List;
+import java.util.Set;
+
+import org.apache.commons.lang3.tuple.ImmutablePair;
+import org.jetbrains.research.kotlinrminer.api.Refactoring;
+import org.jetbrains.research.kotlinrminer.api.RefactoringType;
+import org.jetbrains.research.kotlinrminer.decomposition.AbstractCodeMapping;
+import org.jetbrains.research.kotlinrminer.decomposition.VariableDeclaration;
+import org.jetbrains.research.kotlinrminer.diff.CodeRange;
+import org.jetbrains.research.kotlinrminer.uml.UMLOperation;
+
+public class RenameVariableRefactoring implements Refactoring {
+    private final VariableDeclaration originalVariable;
+    private final VariableDeclaration renamedVariable;
+    private final UMLOperation operationBefore;
+    private final UMLOperation operationAfter;
+    private final Set<AbstractCodeMapping> variableReferences;
+
+    public RenameVariableRefactoring(
+            VariableDeclaration originalVariable,
+            VariableDeclaration renamedVariable,
+            UMLOperation operationBefore,
+            UMLOperation operationAfter,
+            Set<AbstractCodeMapping> variableReferences) {
+        this.originalVariable = originalVariable;
+        this.renamedVariable = renamedVariable;
+        this.operationBefore = operationBefore;
+        this.operationAfter = operationAfter;
+        this.variableReferences = variableReferences;
+    }
+
+    public RefactoringType getRefactoringType() {
+        if (originalVariable.isParameter() && renamedVariable.isParameter())
+            return RefactoringType.RENAME_PARAMETER;
+        if (!originalVariable.isParameter() && renamedVariable.isParameter())
+            return RefactoringType.PARAMETERIZE_VARIABLE;
+        if (!originalVariable.isAttribute() && renamedVariable.isAttribute())
+            return RefactoringType.REPLACE_VARIABLE_WITH_ATTRIBUTE;
+        return RefactoringType.RENAME_VARIABLE;
+    }
+
+    public String getName() {
+        return this.getRefactoringType().getDisplayName();
+    }
+
+    public VariableDeclaration getOriginalVariable() {
+        return originalVariable;
+    }
+
+    public VariableDeclaration getRenamedVariable() {
+        return renamedVariable;
+    }
+
+    public UMLOperation getOperationBefore() {
+        return operationBefore;
+    }
+
+    public UMLOperation getOperationAfter() {
+        return operationAfter;
+    }
+
+    public Set<AbstractCodeMapping> getVariableReferences() {
+        return variableReferences;
+    }
+
+    public String toString() {
+        return getName() + "\t" +
+                originalVariable +
+                " to " +
+                renamedVariable +
+                " in method " +
+                operationAfter +
+                " in class " + operationAfter.getClassName();
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((operationAfter == null) ? 0 : operationAfter.hashCode());
+        result = prime * result + ((operationBefore == null) ? 0 : operationBefore.hashCode());
+        result = prime * result + ((originalVariable == null) ? 0 : originalVariable.hashCode());
+        result = prime * result + ((renamedVariable == null) ? 0 : renamedVariable.hashCode());
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+        RenameVariableRefactoring other = (RenameVariableRefactoring) obj;
+        if (operationAfter == null) {
+            if (other.operationAfter != null)
+                return false;
+        } else if (!operationAfter.equals(other.operationAfter))
+            return false;
+        if (operationBefore == null) {
+            if (other.operationBefore != null)
+                return false;
+        } else if (!operationBefore.equals(other.operationBefore))
+            return false;
+        if (originalVariable == null) {
+            if (other.originalVariable != null)
+                return false;
+        } else if (!originalVariable.equals(other.originalVariable))
+            return false;
+        if (renamedVariable == null) {
+            if (other.renamedVariable != null)
+                return false;
+        } else if (!renamedVariable.equals(other.renamedVariable))
+            return false;
+        return true;
+    }
+
+    public Set<ImmutablePair<String, String>> getInvolvedClassesBeforeRefactoring() {
+        Set<ImmutablePair<String, String>> pairs = new LinkedHashSet<>();
+        pairs.add(new ImmutablePair<>(getOperationBefore().getLocationInfo().getFilePath(),
+                                      getOperationBefore().getClassName()));
+        return pairs;
+    }
+
+    public Set<ImmutablePair<String, String>> getInvolvedClassesAfterRefactoring() {
+        Set<ImmutablePair<String, String>> pairs = new LinkedHashSet<>();
+        pairs.add(
+                new ImmutablePair<>(getOperationAfter().getLocationInfo().getFilePath(),
+                                    getOperationAfter().getClassName()));
+        return pairs;
+    }
+
+    @Override
+    public List<CodeRange> leftSide() {
+        List<CodeRange> ranges = new ArrayList<>();
+        ranges.add(originalVariable.codeRange()
+                           .setDescription("original variable declaration")
+                           .setCodeElement(originalVariable.toString()));
+        return ranges;
+    }
+
+    @Override
+    public List<CodeRange> rightSide() {
+        List<CodeRange> ranges = new ArrayList<>();
+        ranges.add(renamedVariable.codeRange()
+                           .setDescription("renamed variable declaration")
+                           .setCodeElement(renamedVariable.toString()));
+        return ranges;
+    }
+}
+
